refactor: 룰렛 데이터 React-Query로 관리하기
