+++ mechuri/src/app/roulette/page.tsx
@@ -1,20 +1,12 @@
 import React from 'react';
-import Select from '../../components/roulette/Select';
 import NavBar from '@/components/commons/NavBar';
-import Roulette from '../../components/roulette/Roulette';
+import RouletteSubPage from '@/components/roulette/RouletteSubPage';
 
 export default function RoulettePage() {
   return (
     <div className="h-screen">
       <NavBar></NavBar>
-      <div className="grid grid-cols-3 items-center justify-center h-[calc(100vh-4rem)] bg-[#f6f6f6]">
-        <div className="flex justify-center items-center h-3/4">
-          <Select></Select>
-        </div>
-        <div className="col-span-2">
-          <Roulette></Roulette>
-        </div>
-      </div>
+      <RouletteSubPage></RouletteSubPage>
     </div>
   );
 }
